requirements.txt
GIMP with Python Fu Console